requirements.txt
flask
python-dotenv
requests
transformers==4.38.2
torch==2.0.1
sentencepiece
pdfminer.six
